make a login function and delete function again
