--- src/contexts/User.js
@@ -1,34 +1,0 @@
-import { createContext, useState } from 'react';
-
-export const UserContext = createContext();
-
-export const UserProvider = (props) => {
-    const [loggedInUser, setLoggedInUser] = useState([
-        {
-          username: 'butter_bridge',
-          name: 'jonny',
-          avatar_url:
-            'https://www.healthytherapies.com/wp-content/uploads/2016/06/Lime3.jpg'
-        },
-        {
-          username: 'icellusedkars',
-          name: 'sam',
-          avatar_url: 'https://avatars2.githubusercontent.com/u/24604688?s=460&v=4'
-        },
-        {
-          username: 'rogersop',
-          name: 'paul',
-          avatar_url: 'https://avatars2.githubusercontent.com/u/24394918?s=400&v=4'
-        },
-        {
-          username: 'lurker',
-          name: 'do_nothing',
-          avatar_url:
-            'https://www.golenbock.com/wp-content/uploads/2015/01/placeholder-user.png'
-        }
-      ]);
-
-      return <UserContext.Provider value={{ loggedInUser, setLoggedInUser }}>
-        {props.children}
-        </UserContext.Provider>;
-};
