+++ src/contexts/User.jsx
@@ -1,0 +1,11 @@
+import { createContext, useState } from 'react';
+
+export const UserContext = createContext();
+
+export const UserProvider = (props) => {
+    const [loggedInUser, setLoggedInUser] = useState({});
+
+      return <UserContext.Provider value={{ loggedInUser, setLoggedInUser }}>
+        {props.children}
+        </UserContext.Provider>;
+};
